Added demographic graph
Added some new paths to the urls.py.
Added the address field to the account setting.
Revise profile.
Query the database and graph the data inside the profile.
Query the database for demographic data and graph it.
Revise the navigation bar.
Revise the value screen.
Revise the learn more screen.