Added the service score calculation to the views.py
Added the service score calculation to the views.py
Added some new paths to the urls.py.
Revise the value screen.
Revise the views.py.
Revise the learn more screen.
Query the database to obtain the value score of Customer and Representative and graph the data inside the Learn more.
Perform the service score calculation and graph it inside the learn more screen.